refactor: 불필요한 FSD 구조 제거 (#1)
