--- src/apis/index.ts
@@ -1,1 +1,0 @@
-export * from './savingsApi';
