--- src/constants/baseUrl.ts
@@ -1,1 +1,0 @@
-export const API_BASE_URL = '/api';
