+++ src/pages/SavingsCalculatorPage/apis.ts
@@ -1,0 +1,6 @@
+const baseUrl = '/api';
+
+export const api = {
+  savingsProducts: `${baseUrl}/savings-products`,
+  savingsProduct: (id: string) => `${baseUrl}/savings-products/${id}`, // (가정)
+};
