+++ src/pages/SavingsCalculatorPage/components/ProductList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Assets, colors, ListRow } from 'tosslib';
-import { formatAmount } from 'utils';
-import type { SavingsProduct } from '../../../features/savings/types';
+import { formatAmount } from 'pages/SavingsCalculatorPage/utils';
+import type { SavingsProduct } from '../types';
 
 export interface ProductListProps {
   products: SavingsProduct[];
