+++ src/pages/SavingsCalculatorPage/components/index.ts
@@ -1,3 +1,4 @@
+export * from './common';
 export { default as SavingsForm } from './SavingsForm';
 export { default as FilteredProducts } from './FilteredProducts';
 export { default as CalculationResult } from './CalculationResult';
