+++ src/pages/SavingsCalculatorPage/hooks/useProductSelection.ts
