+++ src/pages/SavingsCalculatorPage/types.ts
@@ -1,4 +1,11 @@
-import { SavingsProductAPIResponse } from 'types';
+export interface SavingsProductAPIResponse {
+  id: string;
+  name: string;
+  annualRate: number;
+  minMonthlyAmount: number;
+  maxMonthlyAmount: number;
+  availableTerms: number;
+}
 
 export type SavingsProduct = SavingsProductAPIResponse;
